+++ web3/contract/utils.py
@@ -108,7 +108,7 @@ def call_contract_function(
     output_types = get_abi_output_types(fn_abi)
 
     try:
-        output_data = w3.codec.decode_abi(output_types, return_data)
+        output_data = w3.codec.decode(output_types, return_data)
     except DecodingError as e:
         # Provide a more helpful error message than the one provided by
         # eth-abi-utils
@@ -184,7 +184,7 @@ async def async_call_contract_function(
     output_types = get_abi_output_types(fn_abi)
 
     try:
-        output_data = async_w3.codec.decode_abi(output_types, return_data)
+        output_data = async_w3.codec.decode(output_types, return_data)
     except DecodingError as e:
         # Provide a more helpful error message than the one provided by
         # eth-abi-utils
@@ -218,11 +218,13 @@ async def async_call_contract_function(
 
 async def async_parse_block_identifier(
     w3: "Web3", block_identifier: BlockIdentifier
-) -> Awaitable[BlockIdentifier]:
+) -> BlockIdentifier:
+    if block_identifier is None:
+        return w3.eth.default_block
     if isinstance(block_identifier, int):
         return await async_parse_block_identifier_int(w3, block_identifier)
-    elif block_identifier in ["latest", "earliest", "pending"]:
-        return block_identifier  # type: ignore
+    elif block_identifier in ["latest", "earliest", "pending", "safe", "finalized"]:
+        return block_identifier
     elif isinstance(block_identifier, bytes) or is_hex_encoded_block_hash(
         block_identifier
     ):
@@ -234,7 +236,7 @@ async def async_parse_block_identifier(
 
 async def async_parse_block_identifier_int(
     w3: "Web3", block_identifier_int: int
-) -> Awaitable[BlockNumber]:
+) -> BlockNumber:
     if block_identifier_int >= 0:
         block_num = block_identifier_int
     else:
@@ -243,7 +245,7 @@ async def async_parse_block_identifier_int(
         block_num = last_block_num + block_identifier_int + 1
         if block_num < 0:
             raise BlockNumberOutofRange
-    return BlockNumber(block_num)  # type: ignore
+    return BlockNumber(block_num)
 
 
 def transact_with_contract_function(
